requirements.txt
pyttsx3
sounddevice
soundfile
pywhispercpp
twilio
python-dotenv